requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LangChain e LLMs
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langgraph==0.0.11
openai==1.3.0

# Processamento de Imagem e Vídeo
torch==2.2.1+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.17.1+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
opencv-python-headless==4.8.1.78
Pillow==10.0.1
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
ffmpeg-python==0.2.0

# Processamento de Áudio
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# Armazenamento no GCP
google-cloud-storage==2.13.0
google-cloud-aiplatform==1.36.0

# Utilitários
numpy==1.26.0
pandas==2.1.1
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6

# Segurança e Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Outras dependências
# Atualizado para versões compatíveis
huggingface-hub>=0.16.4,<0.18.0  # Compatível com transformers e sentence-transformers
transformers>=4.35.0,<4.36.0  # Versão específica que funciona bem
sentence-transformers>=2.2.2,<3.0.0  # Versão compatível

# Para suporte a modelos 3D (comentado por problemas de instalação)
# trimesh==3.23.5
# pygltflib==1.15.2
# pyrender==0.1.45
# pymeshlab==2023.12
# open3d==0.17.0
# pytorch3d==0.7.5  # Requer instalação especial
# kaolin==0.13.0
# numpy-quaternion==2023.7.10
# pyvista==0.43.0
# pyvistaqt==0.9.0
# vedo==2023.4.6
# pyntcloud==2023.8.0
# pysdf==0.1.17
# pymcubes==0.1.4
# pyrr==0.10.3
# pymesh2==0.3.13
# scikit-image==0.22.0
# pytorch-lightning==2.1.0

# Para suporte a GPU
# As versões específicas são gerenciadas automaticamente pelo PyTorch
nvidia-cudnn-cu11>=8.6.0.163
nvidia-cublas-cu11>=11.11.3.6
nvidia-cuda-nvrtc-cu11>=11.7.99
nvidia-cuda-runtime-cu11>=11.7.99

# Para monitoramento
opentelemetry-sdk==1.19.0
opentelemetry-exporter-gcp-monitoring>=1.9.0a0  # Atualizado para uma versão estável

# Para testes (opcional)
pytest==7.4.2
pytest-cov==4.1.0

# Para desenvolvimento local (não incluído na imagem de produção)
# ipython
# jupyter
# ipykernel
# black
# isort
# flake8
